Desktop: move settings into a full-page view
Promote the desktop settings surfaces from cramped side panels into a
dedicated full-page view with a left-nav rail: Providers, Models, Web
search, and Appearance. Wire the Models section to the existing
GET/PUT /settings so the global default model is finally configurable
from the app.

Per-chat model selection moves into the message bar as a compact
dropdown (Default + models grouped by provider), replacing the old
per-chat Model side panel. Folders stays a contextual side panel since
it is scoped to a single chat.

Extract the shared settings primitives and the Providers/Web-search
panels into src/settings/, and add ModelsPanel and AppearancePanel
(theme selection lifted out of the sidebar brand toggle).
