+++ crates/openwave-desktop/ui/src/styles.css
@@ -1934,8 +1934,80 @@ button {
 .composer-actions {
   display: flex;
   align-items: center;
-  justify-content: flex-end;
+  justify-content: space-between;
   gap: 0.45rem;
+}
+
+.composer-actions-left {
+  display: flex;
+  align-items: center;
+  min-width: 0;
+}
+
+.composer-actions-right {
+  display: flex;
+  align-items: center;
+  gap: 0.45rem;
+  margin-left: auto;
+}
+
+/* ---------- Model menu (composer) ---------- */
+
+.model-menu-trigger {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.3rem;
+  max-width: 14rem;
+  border: 1px solid var(--line);
+  border-radius: 999px;
+  padding: 0.25rem 0.6rem;
+  background: transparent;
+  color: var(--muted-foreground);
+  font-size: 0.8rem;
+  font-weight: 500;
+  transition: color 120ms ease, background-color 120ms ease,
+    border-color 120ms ease;
+}
+
+.model-menu-trigger:hover:not(:disabled) {
+  color: var(--ink);
+  background: var(--accent);
+}
+
+.model-menu-trigger:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.model-menu-trigger svg {
+  flex: 0 0 auto;
+}
+
+.model-menu-label {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.model-menu-content {
+  min-width: 15rem;
+  max-height: 60vh;
+  overflow-y: auto;
+}
+
+.model-menu-group-label {
+  padding: 0.3rem 0.5rem 0.15rem;
+  color: var(--muted-foreground);
+  font-size: 0.7rem;
+  font-weight: 600;
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
+}
+
+.model-menu-item-label {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
 }
 
 .composer-icon-btn {
@@ -2029,6 +2101,92 @@ button {
   background: var(--bg-elevated);
   overflow-y: auto;
   padding: 1.25rem 1rem;
+}
+
+/* A single settings surface (used full-page and in the folders side panel). */
+.settings-panel {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  max-width: 44rem;
+}
+
+/* ---------- Full-page settings ---------- */
+
+.settings-page {
+  display: flex;
+  min-width: 0;
+  min-height: 0;
+  height: 100%;
+  overflow: hidden;
+}
+
+.settings-nav {
+  display: flex;
+  flex: 0 0 auto;
+  flex-direction: column;
+  gap: 2px;
+  width: 220px;
+  padding: 0.85rem 0.6rem;
+  border-right: 1px solid var(--line);
+  overflow-y: auto;
+}
+
+.settings-nav-list {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  margin-top: 0.35rem;
+}
+
+.settings-back {
+  color: var(--muted-foreground);
+}
+
+.settings-page-content {
+  flex: 1 1 auto;
+  min-width: 0;
+  overflow-y: auto;
+  padding: 1.5rem clamp(1rem, 4vw, 2.5rem);
+}
+
+/* ---------- Appearance ---------- */
+
+.theme-options {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.theme-option {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.45rem;
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  padding: 0.5rem 0.85rem;
+  background: var(--bg-elevated);
+  color: var(--ink);
+  font-size: 0.85rem;
+  font-weight: 500;
+  transition: border-color 120ms ease, background-color 120ms ease;
+}
+
+.theme-option:hover {
+  background: var(--accent);
+}
+
+.theme-option svg {
+  color: var(--muted-foreground);
+}
+
+.theme-option.is-active {
+  border-color: var(--primary);
+  background: var(--accent);
+}
+
+.theme-option.is-active svg {
+  color: var(--ink);
 }
 
 .web-search-state {
@@ -2416,6 +2574,15 @@ button {
     grid-row: 2;
     border-top: 1px solid var(--line);
     border-left: 0;
+  }
+
+  .settings-nav {
+    width: 160px;
+    padding: 0.6rem 0.4rem;
+  }
+
+  .settings-page-content {
+    padding: 1.25rem 1rem;
   }
 
   .messages {
